+++ sdk/compute/arm-compute/src/models/mappers.ts
@@ -9823,6 +9823,127 @@ export const GalleryArtifactSafetyProfileBase: coreClient.CompositeMapper = {
   }
 };
 
+export const ImageVersionSecurityProfile: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ImageVersionSecurityProfile",
+    modelProperties: {
+      uefiSettings: {
+        serializedName: "uefiSettings",
+        type: {
+          name: "Composite",
+          className: "GalleryImageVersionUefiSettings"
+        }
+      }
+    }
+  }
+};
+
+export const GalleryImageVersionUefiSettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "GalleryImageVersionUefiSettings",
+    modelProperties: {
+      signatureTemplateNames: {
+        serializedName: "signatureTemplateNames",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      },
+      additionalSignatures: {
+        serializedName: "additionalSignatures",
+        type: {
+          name: "Composite",
+          className: "UefiKeySignatures"
+        }
+      }
+    }
+  }
+};
+
+export const UefiKeySignatures: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "UefiKeySignatures",
+    modelProperties: {
+      pk: {
+        serializedName: "pk",
+        type: {
+          name: "Composite",
+          className: "UefiKey"
+        }
+      },
+      kek: {
+        serializedName: "kek",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "UefiKey"
+            }
+          }
+        }
+      },
+      db: {
+        serializedName: "db",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "UefiKey"
+            }
+          }
+        }
+      },
+      dbx: {
+        serializedName: "dbx",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "UefiKey"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const UefiKey: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "UefiKey",
+    modelProperties: {
+      type: {
+        serializedName: "type",
+        type: {
+          name: "String"
+        }
+      },
+      value: {
+        serializedName: "value",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
 export const ReplicationStatus: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -10360,6 +10481,50 @@ export const SharedGalleryDiskImage: coreClient.CompositeMapper = {
       },
       hostCaching: {
         serializedName: "hostCaching",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const CommunityGalleryMetadata: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "CommunityGalleryMetadata",
+    modelProperties: {
+      publisherUri: {
+        serializedName: "publisherUri",
+        type: {
+          name: "String"
+        }
+      },
+      publisherContact: {
+        serializedName: "publisherContact",
+        type: {
+          name: "String"
+        }
+      },
+      eula: {
+        serializedName: "eula",
+        type: {
+          name: "String"
+        }
+      },
+      publicNames: {
+        serializedName: "publicNames",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      },
+      privacyStatementUri: {
+        serializedName: "privacyStatementUri",
         type: {
           name: "String"
         }
@@ -14759,6 +14924,13 @@ export const GalleryImageVersion: coreClient.CompositeMapper = {
           className: "GalleryImageVersionSafetyProfile"
         }
       },
+      securityProfile: {
+        serializedName: "properties.securityProfile",
+        type: {
+          name: "Composite",
+          className: "ImageVersionSecurityProfile"
+        }
+      },
       replicationStatus: {
         serializedName: "properties.replicationStatus",
         type: {
@@ -16337,6 +16509,13 @@ export const GalleryImageVersionUpdate: coreClient.CompositeMapper = {
           className: "GalleryImageVersionSafetyProfile"
         }
       },
+      securityProfile: {
+        serializedName: "properties.securityProfile",
+        type: {
+          name: "Composite",
+          className: "ImageVersionSecurityProfile"
+        }
+      },
       replicationStatus: {
         serializedName: "properties.replicationStatus",
         type: {
@@ -16698,7 +16877,27 @@ export const CommunityGallery: coreClient.CompositeMapper = {
     name: "Composite",
     className: "CommunityGallery",
     modelProperties: {
-      ...PirCommunityGalleryResource.type.modelProperties
+      ...PirCommunityGalleryResource.type.modelProperties,
+      disclaimer: {
+        serializedName: "properties.disclaimer",
+        type: {
+          name: "String"
+        }
+      },
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
+        }
+      },
+      communityMetadata: {
+        serializedName: "properties.communityMetadata",
+        type: {
+          name: "Composite",
+          className: "CommunityGalleryMetadata"
+        }
+      }
     }
   }
 };
@@ -16792,6 +16991,19 @@ export const CommunityGalleryImage: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      disclaimer: {
+        serializedName: "properties.disclaimer",
+        type: {
+          name: "String"
+        }
+      },
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
+        }
       }
     }
   }
@@ -16826,6 +17038,19 @@ export const CommunityGalleryImageVersion: coreClient.CompositeMapper = {
         type: {
           name: "Composite",
           className: "SharedGalleryImageVersionStorageProfile"
+        }
+      },
+      disclaimer: {
+        serializedName: "properties.disclaimer",
+        type: {
+          name: "String"
+        }
+      },
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
         }
       }
     }
@@ -16939,7 +17164,15 @@ export const SharedGallery: coreClient.CompositeMapper = {
     name: "Composite",
     className: "SharedGallery",
     modelProperties: {
-      ...PirSharedGalleryResource.type.modelProperties
+      ...PirSharedGalleryResource.type.modelProperties,
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        readOnly: true,
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
+        }
+      }
     }
   }
 };
@@ -17033,6 +17266,13 @@ export const SharedGalleryImage: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
+        }
       }
     }
   }
@@ -17067,6 +17307,13 @@ export const SharedGalleryImageVersion: coreClient.CompositeMapper = {
         type: {
           name: "Composite",
           className: "SharedGalleryImageVersionStorageProfile"
+        }
+      },
+      artifactTags: {
+        serializedName: "properties.artifactTags",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "String" } }
         }
       }
     }